requirements.txt
Pillow>=9.4.0,<=11.0.0
requests==2.31.0
PyMuPDF>=1.22.5,<=1.25.0
